requirements.txt
confluent-kafka==2.4.0
elasticsearch==8.13.1
factory_boy==3.3.0
fastapi==0.111.0
ipdb==0.13.13
ipython==8.23.0
pydantic==2.7.1
pytest==8.1.1
python-dotenv==1.0.1
strawberry-graphql[fastapi,debug-server]==0.229.1
testcontainers==4.6.0
pytest-mock==3.14.0
